requirements.txt
# 核心依赖
opencv-python
numpy
pillow
psutil
PySide6

#打包
pyinstaller

# 可选依赖（用于高级功能）
scipy>=1.7.0
